fix: signup logic
