+++ src/apollo/client.ts
@@ -1,6 +1,6 @@
 import { ApolloClient } from "apollo-client";
 import { InMemoryCache } from "apollo-cache-inmemory";
-import { AUTH_KEY } from "@/modules/home/components/signup/constant";
+import { AUTH_KEY } from "@/modules/signup/constant";
 import { setContext } from "apollo-link-context";
 import { createUploadLink } from "apollo-upload-client";
