+++ src/modules/home/use-home/index.ts
@@ -1,24 +1,17 @@
 import { testData } from "./testData";
-import { computed, reactive, ref } from "vue";
+import { reactive, ref } from "vue";
 import { useUpcomingEvents } from "../api";
-import { useRoute } from "vue-router";
 
 const useHome = () => {
   const state = reactive(testData);
   const showModal = ref(false);
   const { result: upcomingEvents } = useUpcomingEvents();
-  const route = useRoute();
-
-  const isSignup = computed(() => {
-    const signup = route.query.signup;
-    return signup === "1";
-  });
 
   function toggleModal() {
     showModal.value = !showModal.value;
   }
 
-  return { state, showModal, toggleModal, upcomingEvents, isSignup };
+  return { state, showModal, toggleModal, upcomingEvents };
 };
 
 export default useHome;
