+++ src/modules/signup/useSignup.ts
@@ -1,6 +1,6 @@
 import { computed, onMounted, ref, watch } from "vue";
 import { testData } from "@/modules/test/testData";
-import { useRouter } from "vue-router";
+import { useRoute, useRouter } from "vue-router";
 import useUser from "@/modules/authentication";
 import { useUpdateUserInfo } from "./api";
 import { User } from "@/apollo/types";
@@ -12,6 +12,7 @@ const useSignup = () => {
   const test = testData;
   const currentModal = ref(ADDITIONAL_INFO);
   const router = useRouter();
+  const route = useRoute();
   const { hotLogout } = useUser();
   const {
     updateUser,
@@ -33,7 +34,7 @@ const useSignup = () => {
 
   const finishModal = () => {
     toggleModal("");
-    router.push("/");
+    router.push(route.path as string);
   };
 
   const cancelSignup = () => {
@@ -42,7 +43,6 @@ const useSignup = () => {
   };
 
   const updateInfo = (data: User) => {
-    console.log(data, "data");
     updateUser({ input: data });
   };
