+++ src/router/authGuard.ts
@@ -1,4 +1,4 @@
-import { AUTH_KEY } from "@/modules/home/components/signup/constant";
+import { AUTH_KEY } from "@/modules/signup/constant";
 import { NavigationGuardWithThis } from "vue-router";
 
 const authGuard: NavigationGuardWithThis<undefined> = (_, __, next) => {
